### 2023-11-24 #6088 @editorialbot
[PRE REVIEW]: hillmaker: A Python package for occupancy analysis in discrete entity flow systems
**Submitting author:** @misken (<a href="http://orcid.org/0000-0001-8471-9116">Mark Isken</a>)
**Repository:** https://github.com/misken/hillmaker
**Branch with paper.md** (empty if default branch):
**Version:** v0.8.0
