### 2023-11-25 #6088 @misken
I added the missing doi values. All of the ones suggested by the bot were indeed correct. There's one proceedings paper that doesn't appear to have a doi.
