### 2023-12-15 #6088 @misken
Thank you. This is my first submission to JOSS and don't have any reviewer suggestions. It's a pretty straightforward application and reviewers wouldn't need any special background beyond basic applied statistics.
