### 2023-12-15 #6154 @mstimberg
👋🏼 @misken @, @rerickson-usgs, @HLasse this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

These checklists contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. The [first comment in this thread](https://github.com/openjournals/joss-reviews/issues/6154#issue-2044238444) also contains links to the JOSS reviewer guidelines.

We aim for reviews to be completed within 6 weeks at the latest, but of course I understand that you might be away/not available during the holiday period at the end of the year. Please let me know if any of you require significantly more time. We can also use `@editorialbot` (our bot) to set automatic reminders if you know you'll be away for a known period of time.
