### 2024-01-05 c0ccefb
docs: minor updatees to CONTRIBUTING.rst
