### 2024-01-05 f3a49e9
docs: add badges to readme
