### 2024-01-05 misken/hillmaker#73 @HLasse
Add CI for updating docs
It seems like managing the documentation (building and pushing) is currently a manual process. Having an action set up on CI for this would probably make maintenance easier in the long run (and would have the added benefit of being able to add a nice green badge with a link to the documentation in the readme).

Whether you want to add this is up to you, this is just a suggestion.

openjournals/joss-reviews#6154
