### 2024-01-05 misken/hillmaker#75 @HLasse
CI: automatic releases
Another CI tool for you to consider.

I've personally found it very useful to use the commit message conventions from [conventional commits](https://www.conventionalcommits.org/en/v1.0.0/) (essentially just prefacing commit msgs with e.g. `feat:`, `fix:` etc.) along with a [semantic release](https://github.com/python-semantic-release/python-semantic-release) workflow to automatically bump version numbering and create new releases to pypi etc.

It takes a little time to set up and get used to, but having automatic releases for every bug fix or feature can save a lot of time in the long run.

Again, just a suggestion - feel free to disregard.
openjournals/joss-reviews#6154
