### 2024-01-05 #6154 @HLasse
My checklist is complete. The package works well and has very thorough documentation. Tests pass, and all the tutorials work as expected.

I've added a few minor PRs and issues for the authors to consider, but nothing major that should hold back the publication.

I recommend the paper for publication.
