### 2024-01-05 misken/hillmaker#75 @misken
Great suggestion and something that I've read about. These are the kinds of things that I definitely want to do to move this project forward.
