### 2024-01-05 misken/hillmaker#73 @misken
I think I do have this automated but let me describe what I do and you can tell me if there's a better way.

The docs are written in Jupyter notebooks generated with jupyter-book. Due to an [issue](https://github.com/executablebooks/jupyter-book/issues/1348) with how jupyter-book handles conda virtual envs, I run a shell script, `build_docs.sh`,  that changes a key notebook setting in all the notebooks, rebuilds the docs and modifies `docs/_config.yml` so that RTD doesn't try to execute the notebooks when changes are pushed. Then once I push, RTD takes over and builds and publishes the documentation.
