### 2024-01-08 823eabe
removed redundant cd commands
