### 2024-01-09 misken/hillmaker#73 @HLasse
Ah, alright - so RTD watches for changes in the docs and automatically updates? That sounds good!
