### 2024-01-11 #6154 @rerickson-usgs
Thank you for inviting me to review this. I should have read the invite better before accepting.
 I've done occupancy modeling in ecology (that is, imperfect detection of species), which is different than hospital occupancy (or patient use) modeling.

That being said, this package seems well done and I recommend accepting. As a feature request, I would to include the equation within the package and not just the supporting paper. I liked the documentation (like this [file](https://hillmaker.readthedocs.io/en/latest/computing_occupancy.html)). An algorithm flow chart might also help.
Basically, something that would allow me to check the implementation of the numerical methods.
Something similar to https://hillmaker.readthedocs.io/en/latest/example_occupancy_analysis.html#step-2-process-each-scenario-with-hillmaker that explains how the main hillmaker function(s) work.
I also liked this page, https://hillmaker.readthedocs.io/en/latest/analysis_date_range.html, of documentation.

I was not able to install the package, but the problem is my work computer has an SSL intercept and conda is not working for me at the moment.
