### 2024-01-12 #6154 @mstimberg
:wave: @rerickson-usgs Many thanks for your review – if you were mismatched for this review, this would also be my fault as an editor. I have to admit, I wasn't aware of the different uses of the term "occupancy modelling"… Do you feel that you are confident to recommend the paper for acceptance nevertheless? If not, I will find another reviewer that is familiar with the domain.

> I was not able to install the package, but the problem is my work computer has an SSL intercept and conda is not working for me at the moment.

I am afraid that installing and running the package is a necessary condition for a review. If you don't think you'll be able to do so on your work machine, then this would be another reason for me to find another reviewer.  In case it helps: if you can get your IT department to provide you with the certificates it uses, there are (non-trivial…) ways to make conda use them (https://docs.conda.io/projects/conda/en/latest/user-guide/configuration/non-standard-certs.html). An easier solution is to disable SSL verification (https://docs.conda.io/projects/conda/en/latest/user-guide/configuration/disable-ssl-verification.html) but of course this comes with a security risk.

Regardless of your answer to these questions, thanks again for volunteering to review and the time you spent on this :pray:!
