### 2024-01-12 #6154 @rerickson-usgs
@mstimberg  I am trying again on cloud based Jupyter Lab and it's working. It have worked before on my laptop, but just takes a _looooong_ time. Perhaps a warning about slow install times might be helpful.
