### 2024-01-12 #6154 @rerickson-usgs
@mstimberg I was able to complete my review. To put all of my feedback on one comments:

- I could install and the tests passed
- The package is well documented. Some suggestions I have to improve the documentation:
    - [ ] As a feature request, I would to see the equations included within the package and not just the supporting paper. I liked the documentation (like this [file](https://hillmaker.readthedocs.io/en/latest/computing_occupancy.html)).
    - [ ] An algorithm flow chart might also help.  Basically, something that would allow me to check the implementation of the numerical methods. Something similar to https://hillmaker.readthedocs.io/en/latest/example_occupancy_analysis.html#step-2-process-each-scenario-with-hillmaker that explains how the main hillmaker function(s) work.
    - I also really liked this page, https://hillmaker.readthedocs.io/en/latest/analysis_date_range.html, of documentation.
- The code seems well organized. I did not run a style check. But,
- [ ] Please consider adding the style guide you're following to the [CONTRIBUTING file]( https://github.com/misken/hillmaker/blob/main/CONTRIBUTING.rst)
