### 2024-01-12 #6154 @misken
Thanks for your review @rerickson-usgs. Glad you were able to install and run. hillmaker doesn't rely on conda, you can do a pip install and it only has a few dependencies. I've never had the install itself be slow, hmm, but I'll look into that.

Yes, I can certainly add more technical detail in terms of the algorithm to the docs. It's nothing terribly complex. We are in the process of writing a new hillmaker paper for Health Care Management Science and will be writing up all the technical details as part of that. I can adapt some of that for the docs.

I will modify the CONTRIBUTING file and my PyCharm settings to make sure my linter use is consistent and documented.

It's funny about the "occupancy modeling" confusion. When I do lit review related to occupancy modeling, I often get results related to the type of ecological occupancy modeling that you do. In addition, I've actually gotten involved in some ecology research with a stream ecologist colleague at Oakland University doing processing of massive amounts of temperature sensor data from loggers in streams. :)
