### 2024-01-16 #6154 @misken
@rerickson-usgs I added some additional algorithmic details to [the doc page on how occupancy is computed](https://hillmaker.readthedocs.io/en/latest/computing_occupancy.html). Realized that there was a markdown table that wasn't being rendered to the html docs and fixed that - it illustrates the simple occupancy example on that page. I also added a link to the [datetime.py](https://github.com/misken/hillmaker/blob/main/src/hillmaker/bydatetime.py) file that does the heavy lifting on the occupancy computations - it's extremely thoroughly commented. Hopefully this will improve things until we get our next paper written and can use that a additional info on how hillmaker works.
