### 2024-01-16 #6154 @mstimberg
Thanks for the updates @misken. Do these additions address your suggestions/concerns, @rerickson-usgs. In other words, would you recommend the paper/software for acceptance as it is? Thanks :pray: !
