### 2024-01-16 #6154 @rerickson-usgs
@mstimberg Looks good. I like that Python code is easy to read.
