### 2024-01-16 misken/hillmaker#73 @misken
Yep, exactly. Most of the headaches around documentation are just getting JupyterBook and RTD to play nice together.
