### 2024-01-18 3d1f0b4
smaller logo
