+++ src/hillmaker/__init__.py
@@ -2,5 +2,5 @@ from hillmaker.scenario import Scenario, create_scenario
 from hillmaker.legacy import make_hills
 from hillmaker.hills import get_plot, get_summary_df, get_bydatetime_df, get_los_plot, get_los_stats
 
-__version__ = "0.8.0"
+__version__ = "0.8.1"
 
