+++ CONTRIBUTING.rst
@@ -57,16 +57,11 @@ is of course very welcome. In this case, the preferred approach is as follows:
 
      $ python -m pytest
 
-6.  Bump the minor version number (in ``__init__.py``) and make a descriptive
-    entry in ``CHANGELOG.md``. Make changes to the documentation if necessary.
-    New features should be documented by creating a new example, or editing
-    an old one.
-
-7.  Commit your changes and push your branch to GitHub::
+6.  Commit your changes and push your branch to GitHub::
 
      $ git add .
      $ git commit -m "Your detailed description of your changes."
      $ git push origin name-of-your-bugfix-or-feature
 
-8.  Submit a pull request through the GitHub website. The developers may
+7.  Submit a pull request through the GitHub website. The developers may
     suggest changes to the code before the request is ultimately accepted.
