+++ paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 hillmaker is a Python package that computes time of day and day of week specific arrival, departure, and 
 occupancy statistics from 
 transaction data containing arrival and departure timestamps. Typical use is for capacity planning problems in 
-places like hospital emergency departments, surgical recovery rooms or any system in which entities arrive, 
+places like hospital emergency departments, surgical recovery rooms, or any system in which entities arrive, 
 occupy capacity for some amount of time, and then depart. It gets its name from the hill-like nature of 
 summary occupancy plots - see \autoref{fig:occplot}.
 
@@ -38,7 +38,7 @@ summary occupancy plots - see \autoref{fig:occplot}.
 The hillmaker package can be used as a command line application as well as an importable library from Jupyter notebooks
 or Python scripts. There is an object-oriented API
 as well as a function based API. Under the hood, hillmaker relies primarily on numpy [@harris2020array], pandas [@reback2020pandas], 
-matplotlib [@Hunter:2007], seaborn [@Waskom2021] and [Pydantic](https://docs.pydantic.dev/latest/).
+matplotlib [@Hunter:2007], seaborn [@Waskom2021], and [Pydantic](https://docs.pydantic.dev/latest/).
 Transaction data can be in CSV format or a pandas `DataFrame`. The primary outputs of hillmaker are:
 
 - pandas `DataFrame`s and CSV files with arrival, departure, and occupancy summaries by time of day, day of week and, optionally, an entity category.
@@ -56,7 +56,7 @@ use that exhibit significant and important time of day and day of week patterns.
 are related to labor. Staffing and scheduling relies on effectively matching capacity to demand that varies significantly by time of day and day of week. Another motivating problem
 involved developing a surgical patient rerouting plan to temporarily accommodate patients while a post-surgical holding area
 renovation project was underway.  Proper statistical
-analysis of patient arrival, departure, and occupancy patterns, including computation of percentiles, are a critical part of such analyses. Traditional
+analysis of patient arrival, departure, and occupancy patterns, including computation of percentiles, are a critical part of such projects. Traditional
 statistics packages do not include this type of functionality and one-off SQL based approaches are tedious to create
 and generalize. Additionally, the implementation of percentile functions in SQL is dialect dependent. 
 
@@ -76,7 +76,7 @@ the development of hillmaker, it has been used in other domains such as
 [bike share systems](https://bitsofanalytics.org/posts/basic-usage-cycleshare/basic_usage_cycleshare),
 freight operations [@castrellon2023enabling], customer contact centers, and even for analyzing usage patterns of a high performance computing cluster by engineers at a large automobile manufacturer. 
 Any system for which you have data on start and stop times of events, or entry and exit times of entities, is 
-amenable to using hillmaker for characterizing temporal patterns in arrivals, departures, and occupancy (or task starts, task completions and work in progress).
+amenable to using hillmaker for characterizing temporal patterns in arrivals, departures, and occupancy (or task starts, task completions, and work in progress).
 
 Given the rise of Python
 in the scientific computing world, rewriting hillmaker in Python made a lot of sense for its future as an open source
