### 2024-01-26 #6154 @mstimberg
@misken I made some minor changes in a PR (misken/hillmaker#79) – please merge if you are fine with all changes.
